remove scorer
