+++ evaluation/evaluate.py
@@ -9,10 +9,13 @@ from typing import Optional
 import torch
 from datasets import load_dataset
 from fire import Fire
+from infinite_bench.calculate_metrics import calculate_metrics as infinite_bench_scorer
+from loogle.calculate_metrics import calculate_metrics as loogle_scorer
+from ruler.calculate_metrics import calculate_metrics as ruler_scorer
 from tqdm import tqdm
 from transformers import pipeline
+from zero_scrolls.calculate_metrics import calculate_metrics as zero_scrolls_scorer
 
-from infinite_bench.calculate_metrics import calculate_metrics as infinite_bench_scorer
 from kvpress import (
     ExpectedAttentionPress,
     KnormPress,
@@ -21,9 +24,6 @@ from kvpress import (
     SnapKVPress,
     StreamingLLMPress,
 )
-from loogle.calculate_metrics import calculate_metrics as loogle_scorer
-from ruler.calculate_metrics import calculate_metrics as ruler_scorer
-from zero_scrolls.calculate_metrics import calculate_metrics as zero_scrolls_scorer
 
 logger = logging.getLogger(__name__)
