+++ kvpress/presses/observed_attention_press.py
@@ -3,13 +3,12 @@
 
 
 import logging
-from dataclasses import dataclass, field
+from dataclasses import dataclass
 
 import torch
 from torch import nn
 
 from kvpress.presses.scorer_press import ScorerPress
-from kvpress.presses.scorers.observed_attention_scorer import ObservedAttentionScorer
 
 logger = logging.getLogger(__name__)
 
@@ -17,22 +16,38 @@ logger = logging.getLogger(__name__)
 @dataclass
 class ObservedAttentionPress(ScorerPress):
     """
-    This pruner can be used when eager attention is used in the model (i.e. the attention is materialized).
-    It will not return attentions in its output to save memory.
+    The observed attention score is defined as the average attention weight over all prompt tokens
+    Requires output_attentions=True and attn_implementation="eager" to have access to attentions
+    This approach is related to H2O (https://arxiv.org/abs/2306.14048).
     """
 
-    scorer: ObservedAttentionScorer = field(default_factory=ObservedAttentionScorer, init=False)
     compression_ratio: float = 0.0
     output_attentions: bool = False
 
     def __post_init__(self):
-        self.scorer = ObservedAttentionScorer()
         if not self.output_attentions:
             logger.warning(
                 "Model will not return attentions in its output to save memory. Please use DefaultPruner if"
                 " attentions are needed in the output."
             )
         super().__post_init__()
+
+    def score(
+        self,
+        module: nn.Module,
+        hidden_states: torch.Tensor,
+        keys: torch.Tensor,
+        values: torch.Tensor,
+        attentions: torch.Tensor,
+        kwargs,
+    ) -> torch.Tensor:
+        assert attentions is not None, 'Set output_attentions=True and attn_implementation="eager" to use this hook'
+        scores = attentions.sum(2)
+        bsz, num_key_value_heads, n_tokens, _ = keys.shape
+        n_tokens_in_sum = torch.arange(n_tokens, 0, -1).to(attentions.device, attentions.dtype)
+        scores = scores / n_tokens_in_sum
+        scores = scores.view(bsz, num_key_value_heads, -1, n_tokens).mean(2)
+        return scores
 
     def forward_hook(self, module: nn.Module, input: list[torch.Tensor], kwargs: dict, output: list):
         output = super().forward_hook(module, input, kwargs, output)
