+++ kvpress/presses/random_press.py
@@ -1,14 +1,22 @@
-from dataclasses import dataclass, field
+from dataclasses import dataclass
+
+import torch
+from torch import nn
 
 from kvpress.presses.scorer_press import ScorerPress
-from kvpress.presses.scorers.random_scorer import RandomScorer
 
 
 @dataclass
 class RandomPress(ScorerPress):
-    scorer: RandomScorer = field(default_factory=RandomScorer, init=False)
-    compression_ratio: float = 0.0
+    """Randomly prune KV pairs"""
 
-    def __post_init__(self):
-        self.scorer = RandomScorer()
-        super().__post_init__()
+    def score(
+        self,
+        module: nn.Module,
+        hidden_states: torch.Tensor,
+        keys: torch.Tensor,
+        values: torch.Tensor,
+        attentions: torch.Tensor,
+        kwargs,
+    ) -> torch.Tensor:
+        return torch.rand(*keys.shape[:-1]).to(keys.device, keys.dtype)
