--- kvpress/presses/scorers/random_scorer.py
@@ -1,23 +1,0 @@
-# SPDX-FileCopyrightText: Copyright (c) 1993-2024 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
-# SPDX-License-Identifier: Apache-2.0
-
-
-import torch
-from torch import nn
-
-from kvpress.presses.scorers.base_scorer import BaseScorer
-
-
-class RandomScorer(BaseScorer):
-    """Randomly prune KV pairs"""
-
-    def score(
-        self,
-        module: nn.Module,
-        hidden_states: torch.Tensor,
-        keys: torch.Tensor,
-        values: torch.Tensor,
-        attentions: torch.Tensor,
-        kwargs,
-    ) -> torch.Tensor:
-        return torch.rand(*keys.shape[:-1]).to(keys.device, keys.dtype)
