--- kvpress/presses/scorers/streaming_llm_scorer.py
@@ -1,40 +1,0 @@
-# SPDX-FileCopyrightText: Copyright (c) 1993-2024 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
-# SPDX-License-Identifier: Apache-2.0
-
-
-from dataclasses import dataclass
-
-import torch
-from torch import nn
-
-from kvpress.presses.scorers.base_scorer import BaseScorer
-
-
-@dataclass
-class StreamingLLMScorer(BaseScorer):
-    """
-    Prune a fixed number of KV pairs at the beginning and end of the sequence (https://arxiv.org/abs/2309.17453)
-    We keep the first n_sink tokens and the last n_local tokens.
-    n_local is computed using the compression ratio.
-    """
-
-    compression_ratio: float = 0.0
-    n_sink: int = 4
-
-    def score(
-        self,
-        module: nn.Module,
-        hidden_states: torch.Tensor,
-        keys: torch.Tensor,
-        values: torch.Tensor,
-        attentions: torch.Tensor,
-        kwargs,
-    ) -> torch.Tensor:
-
-        q_len = hidden_states.shape[1]
-        assert q_len > self.n_sink, f"Input should contain more tokens than n_sink={self.n_sink}"
-        n_pruned = q_len - int(q_len * (1 - self.compression_ratio))
-        scores = torch.ones_like(keys[..., 0])
-        scores[:, :, self.n_sink : self.n_sink + n_pruned] = 0
-
-        return scores
