+++ kvpress/presses/streaming_llm_press.py
@@ -1,15 +1,36 @@
-from dataclasses import dataclass, field
+from dataclasses import dataclass
+
+import torch
+from torch import nn
 
 from kvpress.presses.scorer_press import ScorerPress
-from kvpress.presses.scorers.streaming_llm_scorer import StreamingLLMScorer
 
 
 @dataclass
 class StreamingLLMPress(ScorerPress):
-    scorer: StreamingLLMScorer = field(default_factory=StreamingLLMScorer, init=False)
+    """
+    Prune a fixed number of KV pairs at the beginning and end of the sequence (https://arxiv.org/abs/2309.17453)
+    We keep the first n_sink tokens and the last n_local tokens.
+    n_local is computed using the compression ratio.
+    """
+
     compression_ratio: float = 0.0
     n_sink: int = 4
 
-    def __post_init__(self):
-        self.scorer = StreamingLLMScorer(n_sink=self.n_sink)
-        super().__post_init__()
+    def score(
+        self,
+        module: nn.Module,
+        hidden_states: torch.Tensor,
+        keys: torch.Tensor,
+        values: torch.Tensor,
+        attentions: torch.Tensor,
+        kwargs,
+    ) -> torch.Tensor:
+
+        q_len = hidden_states.shape[1]
+        assert q_len > self.n_sink, f"Input should contain more tokens than n_sink={self.n_sink}"
+        n_pruned = q_len - int(q_len * (1 - self.compression_ratio))
+        scores = torch.ones_like(keys[..., 0])
+        scores[:, :, self.n_sink : self.n_sink + n_pruned] = 0
+
+        return scores
